feat(go): import encoding/json if need be
AnyOf types in particular need special json un/marshaling
handling, which is done by implementing JSONMarshaler
and JSONUnmarshaler interfaces for the type. Often
these implementations delegate eventual marshaling
of children fields to the standard lib json package.

This provides the document with the necessary import
statement if it appears it needs it.

Signed-off-by: meows <b5c6@protonmail.com>